fix: use float32 for average color computation to avoid uint8 clipping
The average color matrix was cast to uint8 which clips and quantizes
average color values used for hue search and thresholding. Using float32
preserves precision and prevents incorrect hue shifts.

Fixes #20

+++ correct.py
@@ -53,7 +53,7 @@ def get_filter_matrix(mat):
     mat = cv2.resize(mat, (256, 256))
 
     # Get average values of RGB
-    avg_mat = np.array(cv2.mean(mat)[:3], dtype=np.uint8)
+    avg_mat = np.array(cv2.mean(mat)[:3], dtype=np.float32)
     
     # Find hue shift so that average red reaches MIN_AVG_RED
     new_avg_r = avg_mat[0]
